Adds the -debug_mode flag to atac.bds. This is false by default. If it is false, at the end of the script all intermediate BAM files and indices will be cleaned up to save disk space. If it is true, the behavior will be exactly the same as it is now.
Note that this has a potential unwanted side effect: since the intermediate files will be deleted upon a successful run,
rerunning the script will cause all of the tasks to execute again. Ideally we would use BDS' goals and dependencies
features to prevent this from happening, instead of keeping everything at the task level. But that would involve
a larger refactoring of the code.